Collect auth.ldap tests in the same file
